+++ src/lib.rs
@@ -17,7 +17,9 @@ pub mod parser;
 pub mod source;
 
 pub use diagnostic::{Diagnostic, DiagnosticSink, Severity};
-pub use evaluator::{BuiltinFunction, Environment, Evaluator, ModuleValue, Value};
-pub use lexer::{Keyword, Lexer, Token, TokenKind};
-pub use parser::{AstKind, AstNode, BinaryOperator, Parser};
+pub use evaluator::{
+    BuiltinFunction, Environment, Evaluator, ModuleValue, Type, TypeContext, Value,
+};
+pub use lexer::{Keyword, Lexer, Token, TokenKind, TypeName};
+pub use parser::{AstKind, AstNode, BinaryOperator, Parameter, Parser, TypeAnnotation};
 pub use source::{BytePos, SourceFile, Span};